utilizing Pokemon TCG API and fetching data
