+++ backend/routes/searchRoutes.js
@@ -1,8 +1,11 @@
 import express from 'express';
-import searchProducts from '../controllers/productController.js';
+// import searchProducts from '../controllers/productController.js';
+import searchRawCards from '../controllers/rawCardController.js';
 
 const router = express.Router();
 
-router.get("/products/search", searchProducts);
+// router.get("/products/search", searchProducts);
+
+router.get("/cards", searchRawCards);
 
 export default router;
