+++ src/App.jsx
@@ -2,7 +2,8 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import SignInPage from "./pages/auth/SigninPage";
 import SignUpPage from "./pages/auth/SignUpPage";
 import Dashboard from "./pages/Dashboard";
-import SearchResults from "./pages/Product-Search-Page/SearchResults";
+// import SearchResults from "./pages/Product-Search-Page/SearchResults";
+import SearchResultsV2 from "./pages/Product-Search-Page/SearchResultsV2";
 
 
 function App() {
@@ -14,7 +15,7 @@ function App() {
           <Route path="/" element={<SignInPage />} />
           <Route path="/sign-up" element={<SignUpPage  />} />
           <Route path="/dashboard" element={<Dashboard />} />
-          <Route path="/search" element={<SearchResults />} />
+          <Route path="/cards" element={<SearchResultsV2 />} />
         </Routes>
       </BrowserRouter>
