+++ src/pages/Dashboard/index.jsx
@@ -1,18 +1,20 @@
 import PortFolioHeader from "../../shared-components/PortfolioHeader";
-import SearchBar from "../Product-Search-Page/SearchBar";
+// import SearchBar from "../Product-Search-Page/SearchBar";
 import { useState } from "react";
+import SearchBarV2 from "../Product-Search-Page/SearchBarV2";
 
 
 const Dashboard = () => {
 
   const [portfolioName, setPortfolioName ] = useState("")
-  const [amountOfPortfolio, setAmountOfPortfolio  ] = useState("")
+  // const [amountOfPortfolio, setAmountOfPortfolio  ] = useState("")
 
 
   return (
     <>
       <div className="max-sm:!items-center flex flex-col pt-10 pl-10 bg-slate-100 opacity-80 w-full h-screen">
-        <SearchBar />
+        {/* <SearchBar /> */}
+        <SearchBarV2 />
         <PortFolioHeader
         textColor={"slate-500"}
         size="md"
