+++ src/shared-components/Card.jsx
@@ -4,38 +4,27 @@ const Card = (props) => {
   const { product } = props;
   const [qty, setQty] = useState(0);
 
-  const images = [
-    "public/images/brilliantStars.png",
-    "public/images/firstEditionBox.avif",
-    "public/images/celebi.png",
-    "public/images/charizard.png",
-    "public/images/es_etb.png",
-    "public/images/prismatic.png",
-    "public/images/skyridgeBoosterbox.png",
-  ];
+  console.log("product getting passed correctly", product )
 
-  const formatter = new Intl.NumberFormat("en-US", {
-    style: "currency",
-    currency: "USD",
-  });
 
-  const dollarFormatter = formatter.format(product["loose-price"] / 100);
-  const combinedName = product["console-name"] + " " + product["product-name"];
-  const productName = combinedName.split(" ").slice(1).join(" ");
+  const getPrice = (product) => {
+    return (
+      product?.tcgplayer?.prices?.['1stEditionHolofoil']?.market ??
+      product?.tcgplayer?.prices?.['unlimitedHolofoil']?.market ??
+      product?.tcgplayer?.prices?.holofoil?.market ??
+      product?.tcgplayer?.prices?.reverseHolofoil?.market
+    );  
+  }
+  const price = getPrice(product);
 
-  const editedSetName = product["console-name"].split(" ");
-  const setName = editedSetName.slice(1).join(" ");
-
-  const randomIdx = Math.floor(Math.random() * 6);
-  const [randomImg, setRandomImg] = useState("");
-
-  useEffect(() => {
-    setRandomImg(images[randomIdx]);
-  }, []);
+ useEffect(() => {
+  getPrice();
+ })
 
   const increaseQty = () => {
     setQty((prev) => prev + 1);
   };
+  console.log('value of get price', getPrice)
 
   return (
     <>
@@ -43,16 +32,20 @@ const Card = (props) => {
         <div className=" flex flex-col border border-slate-200 w-72 h-96 rounded-lg shadow-lg">
           <div>
             <img
-              src={randomImg}
-              className="w-72 h-56 rounded-t-md object-cover"
+              src={product.images.small}
+              className=" mt-3 w-72 h-48 rounded-t-md object-contain"
             />
           </div>
           <div className="pl-4 mt-2">
-            <div className="font-lato text-xl">{productName}</div>
-            <div>{setName}</div>
+            <div className="font-lato text-xl">{product.name}</div>
+            <div>{product.set.name}</div>
+            <div className="flex">
+            <div>{product.rarity}</div>
+            <div className="pl-2">#{product.number}</div>
+            </div>
           </div>
           <div className="mt-auto pl-4 flex flex-col  font-playfair ">
-            <div>{dollarFormatter}</div>
+            <div>${price}</div>
             <div className="flex justify-between mb-1">
               <div>Qty: {qty}</div>
               <button
